Fixed buffer overflow bug in handle_ipv4_fragment().
In fragment reassembly code, the reassembled buffer length is
calculated by walking the linked list of fragments and stopping at
the first fragment with the "more frags" bit set to 0.  However,
the loop which copied data into the reassembled buffer looped over
all fragments.  If the linked list contains two fragments with "more
frags" set to 0, it could result in buffer overflow because the
size was underestimated.

This fix is to only copy data if it will fit (versus exiting the
copy/free loop after first "no more frags".